backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
graphene-django
strawberry-graphql
fastapi
uvicorn
pydantic
gunicorn
celery
flower
redis
rabbitmq
django-allauth
django-cors-headers
psycopg2-binary
pytest
pytest-django
channels[daphne]
channels-redis
uvicorn

